# went from default cost (12) to 4 for performance